refactor FixedNumberGenerator 클래스를 테스트 디렉토리로 이동
